Massive refactor
The main goal of this is to replace the various FromPacket() and
ToPacket() methods with methods that satisfy the encoding
BinaryMarshaler and BinaryUnmarshaler interfaces.

The main code was nearing 4K lines so it was also a good opporunity
to break things apart into more packages.

Funtionally, nothing has changed.  I hope.